# job-hunt/apps/web/src/lib/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Job-hunt API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/discovery/scans": {
      "post": {
        "tags": [
          "discovery"
        ],
        "summary": "Queue Scan",
        "operationId": "queue_scan_api_v1_discovery_scans_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScanIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Queue Scan Api V1 Discovery Scans Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/discovery/candidate-imports": {
      "post": {
        "tags": [
          "discovery"
        ],
        "summary": "Queue Candidate Import",
        "description": "Queue a user-uploaded export; this endpoint never fetches a source site.",
        "operationId": "queue_candidate_import_api_v1_discovery_candidate_imports_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CandidateImportIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Queue Candidate Import Api V1 Discovery Candidate Imports Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/inbox/{item_id}": {
      "get": {
        "tags": [
          "inbox"
        ],
        "summary": "Get Inbox",
        "operationId": "get_inbox_api_v1_inbox__item_id__get",
        "parameters": [
          {
            "name": "item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Item Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Inbox Api V1 Inbox  Item Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/inbox/{item_id}/actions": {
      "post": {
        "tags": [
          "inbox"
        ],
        "summary": "Act",
        "operationId": "act_api_v1_inbox__item_id__actions_post",
        "parameters": [
          {
            "name": "item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Item Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InboxAction"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Act Api V1 Inbox  Item Id  Actions Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/inbox/{item_id}/handoff": {
      "post": {
        "tags": [
          "inbox"
        ],
        "summary": "Handoff Apply",
        "operationId": "handoff_apply_api_v1_inbox__item_id__handoff_post",
        "parameters": [
          {
            "name": "item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Item Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Handoff Apply Api V1 Inbox  Item Id  Handoff Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/documents": {
      "get": {
        "tags": [
          "documents"
        ],
        "summary": "List All",
        "operationId": "list_all_api_v1_documents_get",
        "parameters": [
          {
            "name": "inbox_item_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Inbox Item Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response List All Api V1 Documents Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/documents/generate": {
      "post": {
        "tags": [
          "documents"
        ],
        "summary": "Generate",
        "operationId": "generate_api_v1_documents_generate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Generate Api V1 Documents Generate Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/outreach/drafts": {
      "get": {
        "tags": [
          "outreach"
        ],
        "summary": "List Drafts",
        "operationId": "list_drafts_api_v1_outreach_drafts_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response List Drafts Api V1 Outreach Drafts Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "outreach"
        ],
        "summary": "Create Draft",
        "operationId": "create_draft_api_v1_outreach_drafts_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DraftIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Create Draft Api V1 Outreach Drafts Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CandidateImportIn": {
        "properties": {
          "metadata": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metadata"
          },
          "jobs": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "maxItems": 100,
            "title": "Jobs"
          },
          "posts": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "maxItems": 100,
            "title": "Posts"
          },
          "titles": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 10,
            "title": "Titles"
          },
          "cities": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 10,
            "title": "Cities"
          }
        },
        "type": "object",
        "title": "CandidateImportIn",
        "description": "Safe subset of the browserless export produced on a candidate device."
      },
...
      "DraftIn": {
        "properties": {
          "inbox_item_id": {
            "type": "string",
            "title": "Inbox Item Id"
          },
          "channel": {
            "type": "string",
            "title": "Channel",
            "default": "email"
          },
          "body": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Body"
          },
          "explicit_outreach_request": {
            "type": "boolean",
            "title": "Explicit Outreach Request",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "inbox_item_id"
        ],
        "title": "DraftIn"
      },
      "GenerateIn": {
        "properties": {
          "inbox_item_id": {
            "type": "string",
            "title": "Inbox Item Id"
          },
          "approval_task_id": {
            "type": "string",
            "title": "Approval Task Id"
          }
        },
        "type": "object",
        "required": [
          "inbox_item_id",
          "approval_task_id"
        ],
        "title": "GenerateIn"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "InboxAction": {
        "properties": {
          "action": {
            "type": "string",
            "title": "Action"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          }
        },
        "type": "object",
        "required": [
          "action"
        ],
        "title": "InboxAction"
      },
...
      "ScanIn": {
        "properties": {
          "connector_id": {
            "type": "string",
            "title": "Connector Id",
            "default": "public_ats_fixture"
          },
          "titles": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Titles"
          },
          "cities": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Cities"
          },
          "work_mode": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Work Mode"
          },
          "saved_search_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Saved Search Name"
          }
        },
        "type": "object",
        "title": "ScanIn"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}